# swaggerV1.json
{"openapi":"3.0.0","paths":{"/v1/clients":{"get":{"operationId":"findAllClient","summary":"Get the paginated list of clients. The default limit is 20 and the default offset is 0.","parameters":[{"name":"name","required":false,"in":"query","description":"Name of client","schema":{"nullable":true,"type":"string"}},{"name":"email","required":false,"in":"query","description":"Email of client","schema":{"nullable":true,"type":"string"}},{"name":"phone","required":false,"in":"query","description":"Phone number of client","schema":{"type":"string"}},{"name":"address","required":false,"in":"query","description":"Address of client","schema":{"type":"string"}},{"name":"city","required":false,"in":"query","description":"City of client","schema":{"type":"string"}},{"name":"state","required":false,"in":"query","description":"State of client","schema":{"type":"string"}},{"name":"zip","required":false,"in":"query","description":"Zip code of client","schema":{"type":"string"}},{"name":"offset","required":false,"in":"query","description":"Default value: 0","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Default value: 20","schema":{"type":"number"}}],"responses":{"200":{"description":"List of clients","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pagination"}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Client"]},"post":{"operationId":"createClient","summary":"Create a new client","parameters":[],"requestBody":{"required":true,"description":"Client data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientRequest"}}}},"responses":{"200":{"description":"Client created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Client"]}},"/v1/clients/{id}":{"get":{"operationId":"findClientById","summary":"Get a client by id","parameters":[{"name":"id","required":true,"in":"path","description":"Client id","schema":{"type":"number"}}],"responses":{"200":{"description":"Client found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Client"]},"patch":{"operationId":"updateClien","summary":"Update a client by id","parameters":[{"name":"id","required":true,"in":"path","description":"Client id","schema":{"type":"number"}}],"requestBody":{"required":true,"description":"Client data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientRequest"}}}},"responses":{"200":{"description":"Client updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Client"]},"delete":{"operationId":"deleteClient","summary":"Delete a client by id","parameters":[{"name":"id","required":true,"in":"path","description":"Client id","schema":{"type":"number"}}],"responses":{"200":{"description":"Client deleted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientDto"}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Client"]}},"/v1/services":{"get":{"operationId":"findAllService","summary":"Get the paginated list of services. The default limit is 20 and the default offset is 0.","parameters":[{"name":"name","required":false,"in":"query","description":"Name of client","schema":{"nullable":true,"type":"string"}},{"name":"offset","required":false,"in":"query","description":"Default value: 0","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Default value: 20","schema":{"type":"number"}}],"responses":{"200":{"description":"List of services","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Pagination"}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Service"]},"post":{"operationId":"createService","summary":"Create a new service","parameters":[],"requestBody":{"required":true,"description":"Service data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceRequest"}}}},"responses":{"200":{"description":"Service created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Service"]}},"/v1/services/{id}":{"get":{"operationId":"findServiceById","summary":"Get a service by id","parameters":[{"name":"id","required":true,"in":"path","description":"Service id","schema":{"type":"number"}}],"responses":{"200":{"description":"Service found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"400":{"description":"Invalid query parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Service"]},"patch":{"operationId":"updateClien","summary":"Update a service by id","parameters":[{"name":"id","required":true,"in":"path","description":"Service id","schema":{"type":"number"}}],"requestBody":{"required":true,"description":"Service data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceRequest"}}}},"responses":{"200":{"description":"Service updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Service"]},"delete":{"operationId":"deleteService","summary":"Delete a service by id","parameters":[{"name":"id","required":true,"in":"path","description":"Service id","schema":{"type":"number"}}],"responses":{"200":{"description":"Service deleted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServiceDto"}}}},"400":{"description":"Invalid parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorFormat"}}}}},"tags":["Service"]}},"/v1":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":""}}}},"/v1/version":{"get":{"operationId":"getVersion","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"Smart Power Electric API","description":"Smart Power Electric API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"AUTH_TOKEN":{"scheme":"Bearer","bearerFormat":"Bearer","description":"Please enter token in following format: Bearer ....JWT....","name":"Authorization","type":"http","in":"Header"}},"schemas":{"Pagination":{"type":"object","properties":{"data":{"description":"Data","type":"array","items":{"type":"object"}},"offset":{"type":"integer","description":"Offset","default":0,"nullable":true},"limit":{"type":"integer","description":"Limit","default":10,"nullable":true},"total":{"type":"integer","description":"Total","default":0}},"required":["data","offset","limit","total"]},"ErrorFormat":{"type":"object","properties":{"statusCode":{"type":"number"},"timestamp":{"format":"date-time","type":"string"},"message":{"type":"string"},"name":{"type":"string"},"path":{"type":"string"},"requestId":{"type":"string"}},"required":["statusCode","timestamp","message","name","path","requestId"]},"CreateClientRequest":{"type":"object","properties":{"name":{"type":"string","example":"John Doe","description":"Name of client"},"email":{"type":"string","example":"Email of client","description":"Email of client"},"phone":{"type":"string","example":"1234567890","description":"Phone number of client"},"address":{"type":"string","example":"123 Main St","description":"Address of client"},"city":{"type":"string","example":"City","description":"City of client"},"state":{"type":"string","example":"State","description":"State of client"},"zip":{"type":"string","example":"12345","description":"Zip code of client"}},"required":["name","email"]},"ClientDto":{"type":"object","properties":{"id":{"type":"integer","example":1,"description":"Id of client"},"name":{"type":"string","example":"John Doe","description":"Name of client"},"email":{"type":"string","example":"Email of client","description":"Email of client"},"phone":{"type":"string","example":"1234567890","description":"Phone number of client"},"address":{"type":"string","example":"123 Main St","description":"Address of client"},"city":{"type":"string","example":"City","description":"City of client"},"state":{"type":"string","example":"State","description":"State of client"},"zip":{"type":"string","example":"12345","description":"Zip code of client"}},"required":["id","name","email"]},"CreateServiceRequest":{"type":"object","properties":{"name":{"type":"string","example":"Service Name","description":"Name of service"},"description":{"type":"string","example":"Service Description","description":"Description of service"},"unitCost":{"type":"number","example":100,"description":"Price of service"}},"required":["name"]},"ServiceDto":{"type":"object","properties":{"id":{"type":"integer","example":1,"description":"Id of service"},"name":{"type":"string","example":"Service Name","description":"Name of service"},"description":{"type":"string","example":"Service Description","description":"Description of service"},"unitCost":{"type":"number","example":100,"description":"Price of service"}},"required":["id","name"]}}}}